dfs and bfs code reorganization
depth_first and breadth_first modules are crated under the iter module.
