+++ src/iter/depth_first/dfs_iter.rs
@@ -1,50 +1,50 @@
-use super::{IterKindCore, QueueElement};
+use crate::iter::{IterKindCore, QueueElement};
 use crate::{
     helpers::N,
     tree::{DefaultMemory, DefaultPinVec},
     TreeVariant,
 };
-use alloc::collections::VecDeque;
+use alloc::{vec, vec::Vec};
 use core::marker::PhantomData;
 use orx_pinned_vec::PinnedVec;
 use orx_self_or::SoM;
 use orx_selfref_col::{MemoryPolicy, NodePtr, SelfRefCol};
 
-/// Breadth first search iterator.
-/// This traversal also known as "level-order" ([wikipedia](https://en.wikipedia.org/wiki/Tree_traversal#Breadth-first_search)).
-pub struct BfsIter<
+/// Depth first search iterator.
+/// This traversal also known as "pre-order" ([wikipedia](https://en.wikipedia.org/wiki/Tree_traversal#Pre-order,_NLR)).
+pub struct DfsIter<
     'a,
     K,
     V,
     M = DefaultMemory<V>,
     P = DefaultPinVec<V>,
-    S = VecDeque<<K as IterKindCore<'a, V, M, P>>::QueueElement>,
+    S = Vec<<K as IterKindCore<'a, V, M, P>>::QueueElement>,
 > where
     K: IterKindCore<'a, V, M, P>,
     V: TreeVariant,
     M: MemoryPolicy<V>,
     P: PinnedVec<N<V>>,
-    S: SoM<VecDeque<K::QueueElement>>,
+    S: SoM<Vec<K::QueueElement>>,
 {
     pub(super) col: &'a SelfRefCol<V, M, P>,
     pub(super) queue: S,
     phantom: PhantomData<K>,
 }
 
-impl<'a, K, V, M, P, S> From<BfsIter<'a, K, V, M, P, S>> for (&'a SelfRefCol<V, M, P>, S)
+impl<'a, K, V, M, P, S> From<DfsIter<'a, K, V, M, P, S>> for (&'a SelfRefCol<V, M, P>, S)
 where
     K: IterKindCore<'a, V, M, P>,
     V: TreeVariant,
     M: MemoryPolicy<V>,
     P: PinnedVec<N<V>>,
-    S: SoM<VecDeque<K::QueueElement>>,
+    S: SoM<Vec<K::QueueElement>>,
 {
-    fn from(value: BfsIter<'a, K, V, M, P, S>) -> Self {
+    fn from(value: DfsIter<'a, K, V, M, P, S>) -> Self {
         (value.col, value.queue)
     }
 }
 
-impl<'a, K, V, M, P> BfsIter<'a, K, V, M, P, VecDeque<K::QueueElement>>
+impl<'a, K, V, M, P> DfsIter<'a, K, V, M, P, Vec<K::QueueElement>>
 where
     K: IterKindCore<'a, V, M, P>,
     V: TreeVariant,
@@ -52,10 +52,8 @@ where
     P: PinnedVec<N<V>>,
 {
     pub(crate) fn new(col: &'a SelfRefCol<V, M, P>, root_ptr: NodePtr<V>) -> Self {
-        let mut queue = VecDeque::new();
-        queue.push_back(<K::QueueElement as QueueElement<V>>::from_root_ptr(
-            root_ptr,
-        ));
+        let root = <K::QueueElement as QueueElement<V>>::from_root_ptr(root_ptr);
+        let queue = vec![root];
         Self {
             col,
             queue,
@@ -64,7 +62,7 @@ where
     }
 }
 
-impl<'a, K, V, M, P> BfsIter<'a, K, V, M, P, &'a mut VecDeque<K::QueueElement>>
+impl<'a, K, V, M, P> DfsIter<'a, K, V, M, P, &'a mut Vec<K::QueueElement>>
 where
     K: IterKindCore<'a, V, M, P>,
     V: TreeVariant,
@@ -74,10 +72,10 @@ where
     pub(crate) fn new_with_queue(
         col: &'a SelfRefCol<V, M, P>,
         root_ptr: NodePtr<V>,
-        queue: &'a mut VecDeque<K::QueueElement>,
+        queue: &'a mut Vec<K::QueueElement>,
     ) -> Self {
         queue.clear();
-        queue.push_back(<K::QueueElement as QueueElement<V>>::from_root_ptr(
+        queue.push(<K::QueueElement as QueueElement<V>>::from_root_ptr(
             root_ptr,
         ));
         Self {
@@ -88,21 +86,20 @@ where
     }
 }
 
-impl<'a, K, V, M, P, S> Iterator for BfsIter<'a, K, V, M, P, S>
+impl<'a, K, V, M, P, S> Iterator for DfsIter<'a, K, V, M, P, S>
 where
     K: IterKindCore<'a, V, M, P>,
     V: TreeVariant + 'a,
     M: MemoryPolicy<V> + 'a,
     P: PinnedVec<N<V>> + 'a,
-    S: SoM<VecDeque<K::QueueElement>>,
+    S: SoM<Vec<K::QueueElement>>,
 {
     type Item = K::YieldElement;
 
     fn next(&mut self) -> Option<Self::Item> {
-        let queue: &mut VecDeque<K::QueueElement> = self.queue.get_mut();
-        queue.pop_front().map(|parent| {
-            let children = K::children(&parent);
-            queue.extend(children);
+        self.queue.get_mut().pop().map(|parent| {
+            let children = K::children_rev(&parent);
+            self.queue.get_mut().extend(children);
             K::element(self.col, &parent)
         })
     }
