+++ src/iter/depth_first/dfs_iter_mut.rs
@@ -1,6 +1,7 @@
-use super::{DfsIter, IterKindCore};
+use super::DfsIter;
 use crate::{
     helpers::N,
+    iter::IterKindCore,
     tree::{DefaultMemory, DefaultPinVec},
     TreeVariant,
 };
