+++ src/iter/depth_first/mod.rs
@@ -1,0 +1,10 @@
+mod dfs;
+mod dfs_iter;
+mod dfs_iter_mut;
+
+pub use dfs::{
+    Dfs, DfsCore, DfsOverData, DfsOverDepthData, DfsOverDepthNode, DfsOverDepthSiblingData,
+    DfsOverDepthSiblingNode, DfsOverNode,
+};
+pub use dfs_iter::DfsIter;
+pub use dfs_iter_mut::DfsIterMut;
