+++ src/iter/mod.rs
@@ -1,23 +1,15 @@
-mod bfs;
-mod bfs_iter;
-mod bfs_iter_mut;
-mod dfs;
-mod dfs_iter;
-mod dfs_iter_mut;
+mod breadth_first;
+mod depth_first;
 mod iter_kinds;
 
-pub use bfs::{
-    Bfs, BfsCore, BfsOverData, BfsOverDepthData, BfsOverDepthNode, BfsOverDepthSiblingData,
-    BfsOverDepthSiblingNode, BfsOverNode,
+pub use breadth_first::{
+    Bfs, BfsCore, BfsIter, BfsIterMut, BfsOverData, BfsOverDepthData, BfsOverDepthNode,
+    BfsOverDepthSiblingData, BfsOverDepthSiblingNode, BfsOverNode,
 };
-pub use bfs_iter::BfsIter;
-pub use bfs_iter_mut::BfsIterMut;
-pub use dfs::{
-    Dfs, DfsCore, DfsOverData, DfsOverDepthData, DfsOverDepthNode, DfsOverDepthSiblingData,
-    DfsOverDepthSiblingNode, DfsOverNode,
+pub use depth_first::{
+    Dfs, DfsCore, DfsIter, DfsIterMut, DfsOverData, DfsOverDepthData, DfsOverDepthNode,
+    DfsOverDepthSiblingData, DfsOverDepthSiblingNode, DfsOverNode,
 };
-pub use dfs_iter::DfsIter;
-pub use dfs_iter_mut::DfsIterMut;
 pub use iter_kinds::{
     IterKindCore, IterMutOver, IterOver, NodeDepthSiblingVal, NodeDepthVal, NodeVal, NodeValue,
     NodeValueData, NodeValueNode, OverData, OverDepthData, OverDepthNode, OverDepthSiblingData,
